requirements.txt
Django==1.10.1
django-braces==1.9.0
django-registration-redux==1.4
Pillow==3.3.1
python-magic==0.4.12
